Remove 'user_roles' helper
