+++ src/ggrc-client/js/components/tree/templates/tree-item-attr.mustache
@@ -131,11 +131,9 @@
     <div class="role tree-title-area">
       <div class="item-data">
         <div class="tree-title-area">
-          {{#user_roles instance parentInstance}}
-            <span class="role" rel="tooltip" data-original-title="{{rolesList}}">
-              {{rolesStr}}
-            </span>
-          {{/user_roles}}
+          <span class="role" rel="tooltip">
+            {{userRoles}}
+          </span>
         </div>
       </div>
     </div>
