+++ src/ggrc-client/js/components/tree/tree-item-attr.js
@@ -3,6 +3,7 @@
  Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
  */
 import {formatDate} from '../../plugins/utils/date-utils';
+import {getUserRoles} from '../../plugins/utils/user-utils';
 import template from './templates/tree-item-attr.mustache';
 
 // attribute names considered "default" and representing a date
@@ -40,6 +41,12 @@ export default can.Component.extend({
         type: String,
         get() {
           return this.getDefaultValue();
+        },
+      },
+      userRoles: {
+        type: String,
+        get() {
+          return getUserRoles(this.attr('instance')).join(', ');
         },
       },
     },
