+++ src/ggrc-client/js/plugins/utils/user-utils.js
@@ -5,6 +5,7 @@
 
 import Person from '../../models/business-models/person';
 import RefreshQueue from '../../models/refresh_queue';
+import {getPageInstance} from './current-page-utils';
 import {notifier} from './notifiers-utils';
 
 function cacheCurrentUser() {
@@ -42,7 +43,32 @@ function getPersonInfo(person) {
   return dfd;
 }
 
+const getUserRoles = (person) => {
+  let parentInstance = getPageInstance();
+
+  let roles = {};
+  let allRoleNames = [];
+
+  _.forEach(GGRC.access_control_roles, (role) => {
+    roles[role.id] = role;
+  });
+
+  if (parentInstance && parentInstance.access_control_list) {
+    allRoleNames = _.uniq(parentInstance.access_control_list.filter((acl) => {
+      return acl.person.id === person.id && acl.ac_role_id in roles;
+    }).map((acl) => {
+      return roles[acl.ac_role_id].name;
+    }));
+  } else {
+    let globalRole = person.system_wide_role === 'No Access' ?
+      'No Role' : person.system_wide_role;
+    allRoleNames = [globalRole];
+  }
+  return allRoleNames;
+};
+
 export {
   cacheCurrentUser,
   getPersonInfo,
+  getUserRoles,
 };
